cdk/requirements.txt
-e .
pytest
aws-cdk.aws-elasticsearch
aws-cdk.aws-kinesisfirehose
